Player Smooth rotation
+ base enemy object prefab (no script yet)
