+++ ZW/Library/StateCache/SceneView/95/95539bf910db0305a6e68c32f7df7a45.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":false,"pivot":{"x":-0.7344980835914612,"y":-3.618967056274414,"z":2.1473703384399416},"rotation":{"x":-0.6293036937713623,"y":-0.05650671571493149,"z":0.045960091054439548,"w":-0.7737466096878052},"size":10.98212718963623,"orthographic":false}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":false,"pivot":{"x":-0.7894246578216553,"y":-4.69835090637207,"z":2.8139214515686037},"rotation":{"x":-0.5294840931892395,"y":0.11924663186073303,"z":-0.07547765225172043,"w":-0.8365059494972229},"size":10.98212718963623,"orthographic":false}
