+++ ZW/UserSettings/Layouts/default-2022.dwlt
@@ -19,7 +19,7 @@ MonoBehaviour:
     width: 1600
     height: 817
   m_ShowMode: 4
-  m_Title: Project
+  m_Title: Inspector
   m_RootView: {fileID: 2}
   m_MinSize: {x: 875, y: 300}
   m_MaxSize: {x: 10000, y: 10000}
@@ -119,7 +119,7 @@ MonoBehaviour:
   m_MinSize: {x: 300, y: 100}
   m_MaxSize: {x: 24288, y: 16192}
   vertical: 0
-  controlID: 41
+  controlID: 101
 --- !u!114 &6
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -414,9 +414,9 @@ MonoBehaviour:
   m_SceneHierarchy:
     m_TreeViewState:
       scrollPos: {x: 0, y: 0}
-      m_SelectedIDs: d0060000
-      m_LastClickedID: 0
-      m_ExpandedIDs: 12fbffff
+      m_SelectedIDs: a65b0000
+      m_LastClickedID: 23462
+      m_ExpandedIDs: 12fbffffa65b0000
       m_RenameOverlay:
         m_UserAcceptedRename: 0
         m_Name: 
@@ -796,9 +796,9 @@ MonoBehaviour:
   m_PlayAudio: 0
   m_AudioPlay: 0
   m_Position:
-    m_Target: {x: -0.7344981, y: -3.618967, z: 2.1473703}
+    m_Target: {x: -0.78942466, y: -4.698351, z: 2.8139215}
     speed: 2
-    m_Value: {x: -0.7344981, y: -3.618967, z: 2.1473703}
+    m_Value: {x: -0.78942466, y: -4.698351, z: 2.8139215}
   m_RenderMode: 0
   m_CameraMode:
     drawMode: 0
@@ -844,9 +844,9 @@ MonoBehaviour:
     m_GridAxis: 1
     m_gridOpacity: 0.5
   m_Rotation:
-    m_Target: {x: -0.6293037, y: -0.056506716, z: 0.04596009, w: -0.7737466}
+    m_Target: {x: -0.5294841, y: 0.11924663, z: -0.07547765, w: -0.83650595}
     speed: 2
-    m_Value: {x: -0.6293037, y: -0.056506716, z: 0.04596009, w: -0.7737466}
+    m_Value: {x: -0.530735, y: 0.119132556, z: -0.07565598, w: -0.8357057}
   m_Size:
     m_Target: 10.982127
     speed: 2
