+++ zw_godot/Scene/Prefabs/enemy3.tscn
@@ -1,0 +1,27 @@
+[gd_scene load_steps=6 format=3 uid="uid://cn5k6axnv0prf"]
+
+[ext_resource type="Material" uid="uid://dli12syjo0a7k" path="res://Visual/Material/enemy.tres" id="1_2gn1t"]
+[ext_resource type="Script" path="res://Script/enemy.cs" id="2_2wr0p"]
+
+[sub_resource type="BoxShape3D" id="BoxShape3D_0o0nx"]
+
+[sub_resource type="BoxMesh" id="BoxMesh_hld2k"]
+
+[sub_resource type="Curve3D" id="Curve3D_ndaa3"]
+
+[node name="CharacterBody3D" type="CharacterBody3D"]
+
+[node name="CollisionShape3D" type="CollisionShape3D" parent="."]
+shape = SubResource("BoxShape3D_0o0nx")
+
+[node name="MeshInstance3D" type="MeshInstance3D" parent="."]
+material_override = ExtResource("1_2gn1t")
+mesh = SubResource("BoxMesh_hld2k")
+
+[node name="Node3D" type="Node3D" parent="."]
+script = ExtResource("2_2wr0p")
+
+[node name="Path3D" type="Path3D" parent="Node3D"]
+curve = SubResource("Curve3D_ndaa3")
+
+[node name="PathFollow3D" type="PathFollow3D" parent="Node3D/Path3D"]
