+++ zw_godot/Scene/Prefabs/ground.tscn
@@ -1,0 +1,22 @@
+[gd_scene load_steps=4 format=3 uid="uid://cunm7mu63qeb4"]
+
+[ext_resource type="Material" uid="uid://dgvs2u6xv5acv" path="res://Visual/Material/ground.tres" id="1_1vpj4"]
+
+[sub_resource type="BoxShape3D" id="BoxShape3D_q4en3"]
+size = Vector3(100, 2, 100)
+
+[sub_resource type="BoxMesh" id="BoxMesh_nghuh"]
+size = Vector3(100, 2, 100)
+
+[node name="Ground" type="StaticBody3D"]
+transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0, -1, 0)
+collision_layer = 4
+collision_mask = 3
+
+[node name="CollisionShape3D" type="CollisionShape3D" parent="."]
+shape = SubResource("BoxShape3D_q4en3")
+
+[node name="MeshInstance3D" type="MeshInstance3D" parent="."]
+layers = 4
+material_override = ExtResource("1_1vpj4")
+mesh = SubResource("BoxMesh_nghuh")
