--- zw_godot/Scene/cube.glb.import
@@ -1,32 +1,0 @@
-[remap]
-
-importer="scene"
-importer_version=1
-type="PackedScene"
-uid="uid://buw0bu2y8c025"
-path="res://.godot/imported/cube.glb-9136eac6cdc00ed80853f983ffc68ade.scn"
-
-[deps]
-
-source_file="res://Scene/cube.glb"
-dest_files=["res://.godot/imported/cube.glb-9136eac6cdc00ed80853f983ffc68ade.scn"]
-
-[params]
-
-nodes/root_type="Node3D"
-nodes/root_name="Scene Root"
-nodes/apply_root_scale=true
-nodes/root_scale=1.0
-meshes/ensure_tangents=true
-meshes/generate_lods=true
-meshes/create_shadow_meshes=true
-meshes/light_baking=1
-meshes/lightmap_texel_size=0.2
-skins/use_named_skins=true
-animation/import=true
-animation/fps=30
-animation/trimming=false
-animation/remove_immutable_tracks=true
-import_script/path=""
-_subresources={}
-gltf/embedded_image_handling=1
