+++ zw_godot/Scene/node_3d.tscn
@@ -1,56 +1,16 @@
-[gd_scene load_steps=9 format=3 uid="uid://dh83ht5tbsbiq"]
+[gd_scene load_steps=6 format=3 uid="uid://cl3iq103dfkf4"]
 
-[ext_resource type="Script" path="res://Script/move_and_attack.cs" id="2_ayt3q"]
-[ext_resource type="Material" uid="uid://c311djjji0ny6" path="res://Visual/Material/player.tres" id="2_hhwjn"]
-[ext_resource type="Material" uid="uid://dgvs2u6xv5acv" path="res://Visual/Material/ground.tres" id="3_as0bh"]
-[ext_resource type="Script" path="res://Scene/CameraScript.cs" id="4_p4sdw"]
-
-[sub_resource type="BoxShape3D" id="BoxShape3D_gjkbs"]
-
-[sub_resource type="BoxMesh" id="BoxMesh_rv6i3"]
-
-[sub_resource type="BoxShape3D" id="BoxShape3D_6l4by"]
-size = Vector3(100, 2, 100)
-
-[sub_resource type="BoxMesh" id="BoxMesh_wj8lc"]
-size = Vector3(100, 2, 100)
+[ext_resource type="PackedScene" uid="uid://xjh6b4khdnpy" path="res://Scene/Prefabs/player.tscn" id="1_g8kmd"]
+[ext_resource type="PackedScene" uid="uid://cunm7mu63qeb4" path="res://Scene/Prefabs/ground.tscn" id="3_d0qx0"]
+[ext_resource type="PackedScene" uid="uid://cn5k6axnv0prf" path="res://Scene/Prefabs/enemy3.tscn" id="4_c7k6s"]
+[ext_resource type="Script" path="res://Script/CameraScript.cs" id="4_o2ktk"]
+[ext_resource type="PackedScene" uid="uid://c7t6gy4h460v4" path="res://Scene/Prefabs/enemy2.tscn" id="5_xjs0t"]
 
 [node name="Main" type="Node3D"]
 
-[node name="Player" type="CharacterBody3D" parent="." node_paths=PackedStringArray("rayCast3D", "camera3D", "ground")]
-transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0.5, 0)
-script = ExtResource("2_ayt3q")
-rayCast3D = NodePath("MoveAndAttack/RayCast3D")
+[node name="Player" parent="." node_paths=PackedStringArray("camera3D", "ground") instance=ExtResource("1_g8kmd")]
 camera3D = NodePath("../CameraScript/MainCamera")
 ground = NodePath("../Ground")
-mouseColliderLayers = 6
-
-[node name="CollisionShape3D" type="CollisionShape3D" parent="Player"]
-shape = SubResource("BoxShape3D_gjkbs")
-
-[node name="MeshInstance3D" type="MeshInstance3D" parent="Player"]
-material_override = ExtResource("2_hhwjn")
-mesh = SubResource("BoxMesh_rv6i3")
-
-[node name="MoveAndAttack" type="Node3D" parent="Player"]
-
-[node name="RayCast3D" type="RayCast3D" parent="Player/MoveAndAttack"]
-transform = Transform3D(-0.3251, 0.478189, -0.815871, 0.585685, -0.575551, -0.570714, -0.742485, -0.663382, -0.0929567, 0, 0, 0)
-target_position = Vector3(0, 0, 3.695)
-collision_mask = 6
-
-[node name="Ground" type="StaticBody3D" parent="."]
-transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0, -1, 0)
-collision_layer = 4
-collision_mask = 3
-
-[node name="CollisionShape3D" type="CollisionShape3D" parent="Ground"]
-shape = SubResource("BoxShape3D_6l4by")
-
-[node name="MeshInstance3D" type="MeshInstance3D" parent="Ground"]
-layers = 4
-material_override = ExtResource("3_as0bh")
-mesh = SubResource("BoxMesh_wj8lc")
 
 [node name="DirectionalLight3D" type="DirectionalLight3D" parent="."]
 transform = Transform3D(-0.688354, -0.512917, 0.512917, 0, 0.707107, 0.707107, -0.725374, 0.48674, -0.48674, 0, 0, 0)
@@ -59,8 +19,16 @@ shadow_enabled = true
 
 [node name="CameraScript" type="Node3D" parent="." node_paths=PackedStringArray("camera3D")]
 transform = Transform3D(-4.37114e-08, 0, 1, 0, 1, 0, -1, 0, -4.37114e-08, 0, 0, 0)
-script = ExtResource("4_p4sdw")
+script = ExtResource("4_o2ktk")
 camera3D = NodePath("MainCamera")
 
 [node name="MainCamera" type="Camera3D" parent="CameraScript"]
-transform = Transform3D(1, 0, 0, 0, 0.173648, 0.984808, 0, -0.984808, 0.173648, 0, 14.5, 4)
+transform = Transform3D(1, 0, 0, 0, 0.173648, 0.984808, 0, -0.984808, 0.173648, -0.691587, 14.3827, 2.5637)
+
+[node name="Ground" parent="." instance=ExtResource("3_d0qx0")]
+
+[node name="CharacterBody3D" parent="." instance=ExtResource("4_c7k6s")]
+transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, -10, 0.5, 0)
+
+[node name="RigidBody3D" parent="." instance=ExtResource("5_xjs0t")]
+transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, -8.5, 0.5, 2)
