+++ zw_godot/Script/enemy.cs
@@ -1,0 +1,15 @@
+using Godot;
+using System;
+
+public partial class enemy : Node3D
+{
+	// Called when the node enters the scene tree for the first time.
+	public override void _Ready()
+	{
+	}
+
+	// Called every frame. 'delta' is the elapsed time since the previous frame.
+	public override void _Process(double delta)
+	{
+	}
+}
